backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.11.0
openpyxl==3.1.2
pandas>=2.1.0
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.8.0
fpdf2==2.7.9
aiofiles==23.2.1
